requirements.txt
streamlit
google-generativeai
pandas
gspread
google-auth>=2.0.0
